moved bin theory to ps_utils
