+++ soopercool/coupling_utils.py
@@ -178,31 +178,6 @@ def compute_couplings(mcm, nmt_binning, transfer=None):
     return bpw_windows, inv_coupling
 
 
-def bin_theory_cls(cls, bpwf):
-    """
-    """
-    fields_theory = {"TT": 0, "EE": 1, "BB": 2, "TE": 3}
-    fields_all = ["TT", "TE", "TB", "ET", "BT", "EE", "EB", "BE", "BB"]
-    nl_th = cls["TT"].shape[0]
-
-    size, n_bins, _, nl = bpwf.shape
-    assert size == 9, "Unexpected number of fields in coupling matrix"
-    assert nl <= nl_th, f"Theory spectrum must contain ell up to {nl}."
-
-    cls_dict = {}
-    for fp in fields_all:
-        if fp in fields_theory:
-            cls_dict[fp] = cls[fp][:nl]
-        else:
-            cls_dict[fp] = np.zeros(nl)
-    cls_vec = np.array([cls_dict[fp] for fp in fields_all])
-
-    clb = np.dot(bpwf.reshape(size*n_bins, size*nl), cls_vec.reshape(size*nl))
-    clb = clb.reshape(size, n_bins)
-
-    return {fp: clb[ifp] for ifp, fp in enumerate(fields_all)}
-
-
 def get_couplings_dict(mcm_dict, nmt_binning,
                        transfer_dict=None,
                        ps_names_and_ftags=None,
